+++ FraudProtection/src/test/java/com/paypal/android/fraudprotection/PayPalDataCollectorUnitTest.kt
@@ -23,8 +23,8 @@ import java.util.UUID
 
 class PayPalDataCollectorUnitTest {
 
-    private val sandboxConfig = CoreConfig("fake-client-id", Environment.SANDBOX)
-    private val liveConfig = CoreConfig("fake-client-id", Environment.LIVE)
+    private val sandboxConfig = CoreConfig("fake-client-id", Environment.SANDBOX, "fake-merchant-id")
+    private val liveConfig = CoreConfig("fake-client-id", Environment.LIVE, "fake-merchant-id")
 
     private lateinit var context: Context