+++ PayPalWebPayments/src/test/java/com/paypal/android/paypalwebpayments/PayPalWebCheckoutClientUnitTest.kt
@@ -56,7 +56,7 @@ class PayPalWebCheckoutClientUnitTest {
 
     @MockK
     private val deviceInspector: DeviceInspector = mockk(relaxed = true)
-    private val coreConfig = CoreConfig("fake-client-id", Environment.SANDBOX)
+    private val coreConfig = CoreConfig("fake-client-id", Environment.SANDBOX, "fake-merchant-id")
     private val urlScheme = "com.example.app"
     private val fakeAppSwitchUrl = "https://paypal.com/vault-app-switch"